refactor(MCP): centralize query text rendering logic and remove redundancy
